requirements.txt
PyGithub==1.16.0
wsgiref==0.1.2
